move IPFS tests to nft
