+++ pallets/ajuna-nft-transfer/src/tests.rs
@@ -21,7 +21,7 @@ use frame_support::{
 };
 use frame_system::pallet_prelude::BlockNumberFor;
 use parity_scale_codec::Encode;
-use sp_runtime::testing::H256;
+use sp_runtime::{testing::H256, DispatchError};
 
 type CollectionConfig =
 	pallet_nfts::CollectionConfig<MockBalance, BlockNumberFor<Test>, MockCollectionId>;
@@ -33,6 +33,295 @@ fn create_collection(organizer: MockAccountId) -> MockCollectionId {
 		&CollectionConfig::default(),
 	)
 	.expect("Should have create contract collection")
+}
+
+mod ipfs {
+	use super::*;
+	use sp_runtime::DispatchError;
+
+	#[test]
+	fn set_service_account_works() {
+		ExtBuilder::default().build().execute_with(|| {
+			assert_eq!(ServiceAccount::<Test>::get(), None);
+			assert_ok!(Nft::set_service_account(RuntimeOrigin::root(), ALICE));
+			assert_eq!(ServiceAccount::<Test>::get(), Some(ALICE));
+			System::assert_last_event(mock::RuntimeEvent::Nft(crate::Event::ServiceAccountSet {
+				service_account: ALICE,
+			}));
+		});
+	}
+
+	#[test]
+	fn set_service_account_rejects_non_root_calls() {
+		ExtBuilder::default().build().execute_with(|| {
+			assert_noop!(
+				Nft::set_service_account(RuntimeOrigin::signed(BOB), ALICE),
+				DispatchError::BadOrigin
+			);
+		});
+	}
+
+	#[test]
+	fn prepare_avatar_works() {
+		let prepare_fee = 999;
+		let initial_balance = prepare_fee + MockExistentialDeposit::get();
+
+		ExtBuilder::default()
+			.balances(&[(ALICE, initial_balance)])
+			.build()
+			.execute_with(|| {
+				let avatar_id = create_avatars(1, ALICE, 1)[0];
+				assert_ok!(AAvatars::set_service_account(RuntimeOrigin::root(), BOB));
+				assert_eq!(Balances::free_balance(ALICE), initial_balance);
+				assert_ok!(AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id));
+				assert_eq!(Balances::free_balance(ALICE), initial_balance - prepare_fee);
+				assert_eq!(Preparation::<Test>::get(avatar_id).unwrap().to_vec(), Vec::<u8>::new());
+				System::assert_last_event(mock::RuntimeEvent::AAvatars(
+					crate::Event::PreparedAvatar { avatar_id },
+				));
+			});
+	}
+
+	#[test]
+	fn prepare_avatar_rejects_forging_trading_and_transferring() {
+		ExtBuilder::default()
+			.seasons(&[(SEASON_ID, Season::default())])
+			.locks(&[(ALICE, SEASON_ID, Locks::all_unlocked())])
+			.build()
+			.execute_with(|| {
+				let avatar_ids = create_avatars(SEASON_ID, ALICE, 5);
+				let avatar_id = avatar_ids[0];
+				assert_ok!(AAvatars::set_service_account(RuntimeOrigin::root(), ALICE));
+				assert_ok!(AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id));
+
+				for extrinsic in [
+					AAvatars::set_price(RuntimeOrigin::signed(ALICE), avatar_id, 1_000),
+					AAvatars::transfer_avatar(RuntimeOrigin::signed(ALICE), BOB, avatar_id),
+					AAvatars::forge(
+						RuntimeOrigin::signed(ALICE),
+						avatar_id,
+						avatar_ids[1..3].to_vec(),
+					),
+				] {
+					assert_noop!(extrinsic, Error::<Test>::AlreadyPrepared);
+				}
+			});
+	}
+
+	#[test]
+	fn prepare_avatar_rejects_unsigned_calls() {
+		ExtBuilder::default().build().execute_with(|| {
+			let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+			assert_noop!(
+				AAvatars::prepare_avatar(RuntimeOrigin::none(), avatar_id),
+				DispatchError::BadOrigin
+			);
+		});
+	}
+
+	#[test]
+	fn prepare_avatar_rejects_unowned_avatars() {
+		ExtBuilder::default()
+			.locks(&[(BOB, SEASON_ID, Locks::all_unlocked())])
+			.build()
+			.execute_with(|| {
+				let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+				assert_noop!(
+					AAvatars::prepare_avatar(RuntimeOrigin::signed(BOB), avatar_id),
+					Error::<Test>::Ownership
+				);
+			});
+	}
+
+	#[test]
+	fn prepare_avatar_rejects_avatars_in_trade() {
+		ExtBuilder::default()
+			.seasons(&[(SEASON_ID, Season::default())])
+			.trade_filters(&[(SEASON_ID, TradeFilters::default())])
+			.locks(&[(ALICE, SEASON_ID, Locks::all_unlocked())])
+			.build()
+			.execute_with(|| {
+				let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+				assert_ok!(AAvatars::set_price(RuntimeOrigin::signed(ALICE), avatar_id, 1));
+				assert_noop!(
+					AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id),
+					Error::<Test>::AvatarInTrade
+				);
+			});
+	}
+
+	#[test]
+	fn prepare_avatar_rejects_when_closed() {
+		ExtBuilder::default().build().execute_with(|| {
+			let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+			GlobalConfigs::<Test>::mutate(|config| config.nft_transfer.open = false);
+			assert_noop!(
+				AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id),
+				Error::<Test>::NftTransferClosed
+			);
+		});
+	}
+
+	#[test]
+	fn prepare_avatar_rejects_locked_avatars() {
+		ExtBuilder::default().build().execute_with(|| {
+			let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+			LockedAvatars::<Test>::insert(avatar_id, ());
+			assert_noop!(
+				AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id),
+				Error::<Test>::AvatarLocked
+			);
+		});
+	}
+
+	#[test]
+	fn prepare_avatar_rejects_already_prepared_avatars() {
+		ExtBuilder::default().build().execute_with(|| {
+			let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+			let ipfs_url = IpfsUrl::try_from(Vec::new()).unwrap();
+			Preparation::<Test>::insert(avatar_id, ipfs_url);
+			assert_noop!(
+				AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id),
+				Error::<Test>::AlreadyPrepared
+			);
+		});
+	}
+
+	#[test]
+	fn prepare_avatar_rejects_insufficient_balance() {
+		ExtBuilder::default()
+			.seasons(&[(SEASON_ID, Season::default().prepare_avatar_fee(333))])
+			.balances(&[(ALICE, MockExistentialDeposit::get()), (BOB, 999_999)])
+			.build()
+			.execute_with(|| {
+				let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+				assert_ok!(AAvatars::set_service_account(RuntimeOrigin::root(), BOB));
+				assert_noop!(
+					AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id),
+					sp_runtime::TokenError::FundsUnavailable
+				);
+			});
+	}
+
+	#[test]
+	fn unprepare_avatar_works() {
+		ExtBuilder::default()
+			.seasons(&[(SEASON_ID, Season::default())])
+			.build()
+			.execute_with(|| {
+				let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+				assert_ok!(AAvatars::set_service_account(RuntimeOrigin::root(), ALICE));
+				assert_ok!(AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id));
+				assert_ok!(AAvatars::unprepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id));
+				assert!(!Preparation::<Test>::contains_key(avatar_id));
+				System::assert_last_event(mock::RuntimeEvent::AAvatars(
+					crate::Event::UnpreparedAvatar { avatar_id },
+				));
+			});
+	}
+
+	#[test]
+	fn unprepare_avatar_rejects_unsigned_calls() {
+		ExtBuilder::default().build().execute_with(|| {
+			assert_noop!(
+				AAvatars::unprepare_avatar(RuntimeOrigin::none(), H256::random()),
+				DispatchError::BadOrigin
+			);
+		});
+	}
+
+	#[test]
+	fn unprepare_avatar_rejects_unowned_avatars() {
+		ExtBuilder::default().build().execute_with(|| {
+			let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+			assert_noop!(
+				AAvatars::unprepare_avatar(RuntimeOrigin::signed(BOB), avatar_id),
+				Error::<Test>::Ownership
+			);
+		});
+	}
+
+	#[test]
+	fn unprepare_avatar_rejects_when_closed() {
+		ExtBuilder::default().build().execute_with(|| {
+			let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+			GlobalConfigs::<Test>::mutate(|config| config.nft_transfer.open = false);
+			assert_noop!(
+				AAvatars::unprepare_avatar(RuntimeOrigin::signed(BOB), avatar_id),
+				Error::<Test>::Ownership
+			);
+		});
+	}
+
+	#[test]
+	fn unprepare_avatar_rejects_unprepared_avatars() {
+		ExtBuilder::default().build().execute_with(|| {
+			let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+			assert_noop!(
+				AAvatars::unprepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id),
+				Error::<Test>::NotPrepared
+			);
+		});
+	}
+
+	#[test]
+	fn prepare_ipfs_works() {
+		ExtBuilder::default()
+			.seasons(&[(SEASON_ID, Season::default())])
+			.build()
+			.execute_with(|| {
+				let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+				assert_ok!(AAvatars::set_service_account(RuntimeOrigin::root(), ALICE));
+				assert_ok!(AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id));
+				ServiceAccount::<Test>::put(BOB);
+
+				let ipfs_url = b"ipfs://{CID}/{optional path to resource}".to_vec();
+				let ipfs_url = IpfsUrl::try_from(ipfs_url).unwrap();
+				assert_ok!(AAvatars::prepare_ipfs(
+					RuntimeOrigin::signed(BOB),
+					avatar_id,
+					ipfs_url.clone()
+				));
+				assert_eq!(Preparation::<Test>::get(avatar_id).unwrap(), ipfs_url);
+				System::assert_last_event(mock::RuntimeEvent::AAvatars(
+					crate::Event::PreparedIpfsUrl { url: ipfs_url },
+				));
+
+				let ipfs_url = b"ipfs://123".to_vec();
+				let ipfs_url = IpfsUrl::try_from(ipfs_url).unwrap();
+				assert_ok!(AAvatars::prepare_ipfs(
+					RuntimeOrigin::signed(BOB),
+					avatar_id,
+					ipfs_url.clone()
+				));
+				assert_eq!(Preparation::<Test>::get(avatar_id).unwrap(), ipfs_url);
+				System::assert_last_event(mock::RuntimeEvent::AAvatars(
+					crate::Event::PreparedIpfsUrl { url: ipfs_url },
+				));
+			});
+	}
+
+	#[test]
+	fn prepare_ipfs_rejects_empty_url() {
+		ExtBuilder::default()
+			.seasons(&[(SEASON_ID, Season::default())])
+			.build()
+			.execute_with(|| {
+				let avatar_id = create_avatars(SEASON_ID, ALICE, 1)[0];
+				assert_ok!(AAvatars::set_service_account(RuntimeOrigin::root(), ALICE));
+				assert_ok!(AAvatars::prepare_avatar(RuntimeOrigin::signed(ALICE), avatar_id));
+				ServiceAccount::<Test>::put(BOB);
+
+				assert_noop!(
+					AAvatars::prepare_ipfs(
+						RuntimeOrigin::signed(BOB),
+						avatar_id,
+						IpfsUrl::default()
+					),
+					Error::<Test>::EmptyIpfsUrl
+				);
+			});
+	}
 }
 
 mod store_as_nft {
